+++ changelog.md
@@ -6,6 +6,7 @@
 
 * `[GUI]` Added rich text parsing for full text styling needs
 * `[Audio]` New volume buttons features
+* `[settings]` Add basic bluetooth options window.
 
 ### Other